requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.7
cvzone==1.6.1
cycler==0.12.1
databricks-sdk==0.65.0
dataloader==2.0
datasets==4.0.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
docker==7.1.0
evaluate==0.4.5
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
influx==0.0.1
influxdb==5.3.2
influxdb-client==1.48.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.3.0
nest-asyncio==1.6.0
networkx==3.5
neurokit2==0.2.12
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
packaging==24.2
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotly==6.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
pyshimmer==0.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.2
pyzmq==26.4.0
reactivex==4.0.4
regex==2025.9.1
requests==2.32.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.2
sentencepiece==0.2.0
serial==0.0.97
setuptools==79.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
